mosquito_model/requirements.txt
click==7.1.2
tqdm==4.56.0
pandas==1.2.1
geopandas==0.8.1
# gdal==3.1.2               # via ada_tools (setup.py)  # installed separately
Fiona==1.8.13.post1
rasterio==1.1.0
numpy==1.19.2
country_bounding_boxes==0.2.3
earthengine-api==0.1.256
geetools==0.6.7
python_dateutil==2.8.1
